requirements.txt
Pillow~=9.2.0
bitarray~=2.6.1
